requirements.txt
academictorrents
albumentations==1.4.8
einops
ftfy
imageio==2.32.0
imageio-ffmpeg
kornia==0.7.2
more_itertools==10.3.0
natsort==8.4.0
omegaconf
open_clip_torch==2.20.0
opencv_python
opencv_python_headless
packaging
pudb==2024.1.2
pyparsing==3.1.2
pytest==8.3.2
pytorch_lightning==1.6.5
PyYAML==6.0.1
regex
Requests==2.32.3
scikit_learn==1.4.2
scipy==1.10.1
setuptools
scikit-image
streamlit==1.34.0
sympy
test-tube>=0.7.5
timm==0.9.10
tokenizers==0.19.1
torchmetrics
tqdm==4.65.0
transformers==4.41.1
typing_extensions
triton
matplotlib
wandb
pillow
numpy
taming-transformers-rom1504
git+https://github.com/openai/CLIP.git@main#egg=clip
